[refactor/#17] user-service 시큐리티 기능 게이트웨이로 이동, 커스텀예외 삭제
