--- user-service/src/main/java/hanium/user_service/exception/CustomException.java
@@ -1,10 +1,0 @@
-package hanium.user_service.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CustomException extends RuntimeException {
-    ErrorCode errorCode;
-}
