--- user-service/src/main/java/hanium/user_service/exception/CustomExceptionHandler.java
@@ -1,14 +1,0 @@
-package hanium.user_service.exception;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice // 전역에서 발생 가능한 예외를 처리
-public class CustomExceptionHandler {
-
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<ErrorResponseEntity> handleCustomException(CustomException e) {
-        return ErrorResponseEntity.toResponseEntity(e.getErrorCode());
-    }
-}
