+++ user-service/src/main/java/hanium/user_service/security/SecurityConfig.java
@@ -1,7 +1,5 @@
-package hanium.user_service.config;
+package hanium.user_service.security;
 
-import hanium.user_service.security.JwtAuthenticationFilter;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,14 +9,11 @@ import org.springframework.security.config.annotation.web.configurers.AbstractHt
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 @EnableMethodSecurity(prePostEnabled = true)
 @RequiredArgsConstructor
 @Configuration
 public class SecurityConfig {
-
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -31,14 +26,9 @@ public class SecurityConfig {
                 // 인증, 인가가 필요한 url 지정
                 .authorizeHttpRequests(auth -> auth
                         // permitAll()로 지정된 url은 인증 없이도 요청 허용 - 로그인, 회원가입
-                        .requestMatchers("/api/auth/login", "/api/auth/signup","/error","/api/members/health-check").permitAll()
+                        .requestMatchers("/error", "/api/members/health-check").permitAll()
                         // 나머지 요청은 인증 필요
                         .anyRequest().authenticated()
-                )
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
-                .exceptionHandling(e -> e
-                        .authenticationEntryPoint((request, response, authException) ->
-                                response.sendError(HttpServletResponse.SC_UNAUTHORIZED, authException.getMessage()))
                 )
                 .build();
     }
